requirements.txt
pyserial
matplotlib
pandas
seaborn
reportlab
numpy
